fix: 统一 percent_error 量纲为 int[-5000, 5000]
- 三个处理器输出统一为 int[-5000, 5000]，消除 float[-1,1]→int 无意义压缩
- ring_discovery_processor._no_target() 修复 float 0.0→0 防 AttributeError
- VisualContext 类型注解 float→int
- _handle_track_result 移除 int(x*5000) 转换，直接传递
- 更新 protocol.md / TODO.md / 处理器 docstring

+++ context/visual_state_machine.py
@@ -30,8 +30,8 @@ class VisualContext:
     consecutive_lost_frames: int = 0
 
     state_entry_time: float = 0.0
-    percent_error_x: float = 0.0
-    percent_error_y: float = 0.0
+    percent_error_x: int = 0
+    percent_error_y: int = 0
 
     custom: Dict[str, Any] = field(default_factory=dict)
